docs(samples) - streamliine and clarify webhook samples (#264)
* streamline and clarify webhook samples


* remove sessionInfo config


* 🦉 Updates from OwlBot post-processor

See https://github.com/googleapis/repo-automation-bots/blob/main/packages/owl-bot/README.md
