+++ dialogflow-cx/webhooks.js
@@ -16,8 +16,6 @@
 
 // [START dialogflow_cx_webhook]
 
-// TODO: change entry point to handleWebhook in cloud function
-
 exports.handleWebhook = (request, response) => {
   const tag = request.body.fulfillmentInfo.tag;
   let text = '';
